Module cargo_make::types

types

Defines the various types and aliases used by cargo-make.

Structs

CliArgs
Holds CLI args

Config
Holds the entire configuration such as task definitions and env vars

ConfigSection
Holds the configuration found in the makefile toml config section.

CrateInfo
Holds crate information loaded from the Cargo.toml file.

EnvInfo
Holds env information

ExecutionPlan
Execution plan which defines all steps to run and the order to run them

ExternalConfig
Holds the entire externally read configuration such as task definitions and env vars where all values are optional

FlowInfo
Holds flow information

GitInfo
Holds git info for the given repo directory

PackageInfo
Holds crate package information loaded from the Cargo.toml file package section.

PlatformOverrideTask
Holds a single task configuration for a specific platform as an override of another task

RustInfo
Holds rust info for the current runtime

Step
Execution plan step to execute

Task
Holds a single task configuration such as command and dependencies list

TaskCondition
Holds condition attributes

Workspace
Holds crate workspace info.

Enums

RustChannel
Rust channel type

Functions

get_platform_name
Returns the platform name