Struct cargo_make::types::ConfigSection

pub struct ConfigSection {
    pub skip_core_tasks: Option<bool>,
    pub init_task: Option<String>,
    pub end_task: Option<String>,
    pub load_script: Option<Vec<String>>,
    pub linux_load_script: Option<Vec<String>>,
    pub windows_load_script: Option<Vec<String>>,
    pub mac_load_script: Option<Vec<String>>,
}

Holds the configuration found in the makefile toml config section.

Fields
skip_core_tasks: Option<bool>

If true, the default core tasks will not be loaded

init_task: Option<String>

Init task name which will be invoked at the start of every run

end_task: Option<String>

End task name which will be invoked at the end of every run

load_script: Option<Vec<String>>

Invoked while loading the descriptor file but before loading any extended descriptor

linux_load_script: Option<Vec<String>>

acts like load_script if runtime OS is Linux (takes precedence over load_script)

windows_load_script: Option<Vec<String>>

acts like load_script if runtime OS is Windows (takes precedence over load_script)

mac_load_script: Option<Vec<String>>

acts like load_script if runtime OS is Mac (takes precedence over load_script)

Methods

impl ConfigSection

pub fn new() -> ConfigSection

Creates and returns a new instance.

pub fn extend(self: &mut ConfigSection, extended: &mut ConfigSection)

Copies values from the config section into self.

Arguments

task - The task to copy from

pub fn get_load_script(self: &ConfigSection) -> Option<Vec<String>>

Returns the load script based on the current platform