Struct cargo_make::types::ExternalConfig

pub struct ExternalConfig {
    pub extend: Option<String>,
    pub config: Option<ConfigSection>,
    pub env: Option<HashMap<String, EnvValue>>,
    pub tasks: Option<HashMap<String, Task>>,
}

Holds the entire externally read configuration such as task definitions and env vars where all values are optional

Fields
extend: Option<String>

Path to another toml file to extend

config: Option<ConfigSection>

Runtime config

env: Option<HashMap<String, EnvValue>>

The env vars to setup before running the tasks

tasks: Option<HashMap<String, Task>>

All task definitions

Methods

impl ExternalConfig

pub fn new() -> ExternalConfig

Creates and returns a new instance.