Struct cargo_make::types::Task

pub struct Task {
    pub description: Option<String>,
    pub disabled: Option<bool>,
    pub condition: Option<TaskCondition>,
    pub condition_script: Option<Vec<String>>,
    pub force: Option<bool>,
    pub alias: Option<String>,
    pub linux_alias: Option<String>,
    pub windows_alias: Option<String>,
    pub mac_alias: Option<String>,
    pub install_crate: Option<String>,
    pub install_script: Option<Vec<String>>,
    pub command: Option<String>,
    pub args: Option<Vec<String>>,
    pub script: Option<Vec<String>>,
    pub script_runner: Option<String>,
    pub run_task: Option<String>,
    pub dependencies: Option<Vec<String>>,
    pub linux: Option<PlatformOverrideTask>,
    pub windows: Option<PlatformOverrideTask>,
    pub mac: Option<PlatformOverrideTask>,
}

Holds a single task configuration such as command and dependencies list

Fields
description: Option<String>

Task description

disabled: Option<bool>

if true, the command/script of this task will not be invoked, dependencies however will be

condition: Option<TaskCondition>

if provided all condition values must be met in order for the task to be invoked (will not stop dependencies)

condition_script: Option<Vec<String>>

if script exit code is not 0, the command/script of this task will not be invoked, dependencies however will be

force: Option<bool>

if true, any error while executing the task will be printed but will not break the build

alias: Option<String>

if defined, task points to another task and all other properties are ignored

linux_alias: Option<String>

acts like alias if runtime OS is Linux (takes precedence over alias)

windows_alias: Option<String>

acts like alias if runtime OS is Windows (takes precedence over alias)

mac_alias: Option<String>

acts like alias if runtime OS is Mac (takes precedence over alias)

install_crate: Option<String>

if defined, the provided crate will be installed (if needed) before running the task

install_script: Option<Vec<String>>

if defined, the provided script will be executed before running the task

command: Option<String>

The command to execute

args: Option<Vec<String>>

The command args

script: Option<Vec<String>>

If command is not defined, and script is defined, the provided script will be executed

script_runner: Option<String>

The script runner (defaults to cmd in windows and sh for other platforms)

run_task: Option<String>

The task name to execute

dependencies: Option<Vec<String>>

A list of tasks to execute before this task

linux: Option<PlatformOverrideTask>

override task if runtime OS is Linux (takes precedence over alias)

windows: Option<PlatformOverrideTask>

override task if runtime OS is Windows (takes precedence over alias)

mac: Option<PlatformOverrideTask>

override task if runtime OS is Mac (takes precedence over alias)

Methods

impl Task

fn new() -> Task

Creates and returns a new instance.

fn extend(self: &mut Task, task: &Task)

Copies values from the task into self.

Arguments

task - The task to copy from

fn is_force(self: &Task) -> bool

Returns true if the task force attribute is defined and true

fn get_normalized_task(self: &mut Task) -> Task

Returns a new task based on the override information and current platform.

fn get_alias(self: &Task) -> Option<String>

Returns the alias value based on the current platform and task definition.