Struct cargo_make::types::PlatformOverrideTask

pub struct PlatformOverrideTask {
    pub clear: Option<bool>,
    pub disabled: Option<bool>,
    pub condition: Option<TaskCondition>,
    pub condition_script: Option<Vec<String>>,
    pub force: Option<bool>,
    pub env: Option<HashMap<String, EnvValue>>,
    pub cwd: Option<String>,
    pub install_crate: Option<String>,
    pub install_crate_args: Option<Vec<String>>,
    pub install_script: Option<Vec<String>>,
    pub command: Option<String>,
    pub args: Option<Vec<String>>,
    pub script: Option<Vec<String>>,
    pub script_runner: Option<String>,
    pub run_task: Option<String>,
    pub rust_script: Option<Vec<String>>,
    pub dependencies: Option<Vec<String>>,
}

Holds a single task configuration for a specific platform as an override of another task

Fields
clear: Option<bool>

if true, it should ignore all data in base task

disabled: Option<bool>

if true, the command/script of this task will not be invoked, dependencies however will be

condition: Option<TaskCondition>

if provided all condition values must be met in order for the task to be invoked (will not stop dependencies)

condition_script: Option<Vec<String>>

if script exit code is not 0, the command/script of this task will not be invoked, dependencies however will be

force: Option<bool>

if true, any error while executing the task will be printed but will not break the build

env: Option<HashMap<String, EnvValue>>

The env vars to setup before running the task commands

cwd: Option<String>

The working directory for the task to execute its command/script

install_crate: Option<String>

if defined, the provided crate will be installed (if needed) before running the task

install_crate_args: Option<Vec<String>>

additional cargo install arguments

install_script: Option<Vec<String>>

if defined, the provided script will be executed before running the task

command: Option<String>

The command to execute

args: Option<Vec<String>>

The command args

script: Option<Vec<String>>

If command is not defined, and script is defined, the provided script will be executed

script_runner: Option<String>

The script runner (defaults to cmd in windows and sh for other platforms)

run_task: Option<String>

The task name to execute

rust_script: Option<Vec<String>>

The rust code to execute

dependencies: Option<Vec<String>>

A list of tasks to execute before this task

Methods

impl PlatformOverrideTask

pub fn extend(self: &mut PlatformOverrideTask, task: &mut Task)

Copies values from the task into self.

Arguments

task - The task to copy from